+++ pkg/cluster/internal/create/actions/createworker/gcp.go
@@ -56,7 +56,7 @@ func newGCPBuilder() *GCPBuilder {
 func (b *GCPBuilder) setCapx(managed bool) {
 	b.capxProvider = "gcp"
 	b.capxVersion = "v1.6.1"
-	b.capxImageVersion = "1.6.1-0.1"
+	b.capxImageVersion = "1.6.1-0.1.0-96a7243"
 	b.capxName = "capg"
 	b.capxManaged = managed
 	b.csiNamespace = "kube-system"