+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.17.0-0.4.1 (upcoming)
 
+* [Core] Fix panic when keos_version is not defined
 * [Core] Update runc golang module to fix GHSA-xr7r-f8xq-vfvv
 
 ## 0.17.0-0.4.0 (2024-03-06)